requirements.txt
Django==1.8.3
pillow
django-ckeditor
django-extensions==1.4.9
djangorestframework==3.1.1
django-filter
pytz


